Add ModifyState to ControlFlag
Allows the user to modify the state via the ControlFlag which is returned by the event function and the solout function. The solout function requires the user to find the point and event function will use a root finder to find the point of the crossing. Either can be chosen some cases will require solout as it has extra information such as t_prev to t_curr states via interpolation.