Fix RNG usage in SDE #2
Unknowingly the &self referencing self.rng in the examples required a .clone because it should be a &mut. the noise method of the SDE trait was updated to take a &mut. Previously a RefCell could have worked instead but this required extra knowledge that isn't idiomatic to the users and I was also unaware of this problem.