Started implementing DormandPrince variant of ERK still has issues.
Needs interpolation to be fixed, currently just using dop853 version for debugging and no generalized yet. And the result is slightly incorrect so there is an error somewhere unknown at this time. I believe it has to do with error estimation, step size choice or interpolation as preliminary test of primary stages have shown the same result.